+++ python/tvm/relax/vm.py
@@ -501,7 +501,6 @@ def build(
 
     passes = [relax.transform.ToNonDataflow()]
     passes.append(relax.transform.CallTIRRewrite())
-    passes.append(relax.transform.NaivePlanMemory())
     passes.append(relax.transform.VMMemoryLower())
     passes.append(relax.transform.VMShapeLower())
     seq = tvm.transform.Sequential(passes)
